cpsw_bufq.cc: BUGFIX - spuriously lost wakeup from getSlot
putSlot would only awaken sleepers if the number of available
slots went from 0->1. However, if multiple threads are sleeping
in getSlot and putSlot is called faster than awakening threads
consume the new slots then we are in a situation where the slot
count is > 0 and incrementing without waking more sleepers.