Open/close address objects only on adapters that read/write
This makes the python wrapper simpler. Due to the unknown
life-time of python objects we can in particular not leave
streams open while unattended.

With this modification, a stream endpoint can still be
opened as a IVal_Base and basic properties can be accessed
w/o opening the stream (which now only happens when IStream
is created).

This way, a python wrapper can hold a shared_ptr[IVal_Base]
for basic access while __enter__/__exit__ create and
destroy a stream for reading and writing from a 'with' block.