yaml_cpp python wrapper: work around yaml-cpp assignment weirdness.
Yaml-cpp (0.6.2) has some weird operator=() semantics. It seems that
a 'YAML::Node' behaves a little bit like a 'pointer' to the contents
stored in the node. And assigment seems to work like this (written
as if nodes were pointers):

if ( destnode == NULL /* undefined */ )
  destnode = sourcenode
else
  *destnode = *sourcenode

E.g., if there are two nodes

Node a; Node b;

and we want to assign them from an interator

 a = iter->first;
 b = iter->second;

then that works just fine. Cython, however, introduces
a temporary variable

Node tmp;

 tmp = iter->first;
 a   = tmp;
 tmp = iter->second;
 b   = tmp;

and in this scenario all the nodes b, tmp AND a end up referring
to iter->second.

To work around this and prevent cython from generating an intermediate
(unfortunately it does not help to 'not-declare' operator= in the 'pxd'
file -- cython still assumes an operator= from Node to Node exists
and will use it all over the place) we wrap the iterator dereferencing
and assignment to 'a' and 'b' in a 'handleIterator' wrapper.
Sucks...