yaml_cpp: rework to fix exception thrown when looking up non-existent key
The boost implementation returned an 'undefined' node when a key
didnt' exist. However, the weird assignment semantics of yaml-cpp
don't allow for undefined nodes to be assigned (this raised the
exception).

Node x = map["nonexistent"]
return x;

works fine and returns an 'undefined' node.

However,

Node x;
  x = map["nonexistent"];

throws!!!

Added a check so that we return a Null node when a key is
not found (compatible -- sort of --) with the previous version...